Moves components into files
