+++ src/components/ballots/sections/index.tsx
@@ -1,0 +1,5 @@
+import Candidate from "./candidate"
+import PartyHeader from "./party-header"
+import SectionHeader from "./section-header"
+
+export { Candidate, PartyHeader, SectionHeader }
