+++ src/components/ballots/sections/section-header.tsx
@@ -1,0 +1,9 @@
+import React from "react"
+
+export default function SectionHeader({ ocrResult }: { ocrResult: string }) {
+  return (
+    <p className="font-semibold text-center whitespace-pre-line border">
+      {ocrResult}
+    </p>
+  )
+}
